pyproject.toml
[tool.poetry]
name = "item-rec"
version = "0.0.1"
description = ""
authors = ["BabkenBrs <brsikyanbabken99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
pandas = "^2.1.2"
scipy = "^1.11.3"
matplotlib = "^3.8.0"
tqdm = "^4.66.1"
implicit = "^0.7.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
pylint = "^3.0.2"
black = "^23.10.1"
ipykernel = "^6.26.0"
hydra-core = "^1.3.2"

[tool.ruff]
# Enable all `pydocstyle` rules, limiting to those that adhere to the
# Google convention via `convention = "google"`, below.
select = ["D"]

# On top of the Google convention, disable `D417`, which requires
# documentation for every function parameter.
ignore = ["D417"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
